wip: use programmatic checks in ai agent loop
