+++ packages/@n8n/ai-workflow-builder.ee/evaluations/utils/is-tool.ts
@@ -1,5 +1,1 @@
-import type { INodeTypeDescription } from 'n8n-workflow';
-
-export function isTool(nodeType: INodeTypeDescription): boolean {
-	return nodeType.codex?.subcategories?.AI?.includes('Tools') ?? false;
-}
+export { isTool } from '../../src/programmatic/utils/is-tool';
