+++ packages/@n8n/ai-workflow-builder.ee/src/programmatic/evaluators/index.ts
@@ -1,0 +1,11 @@
+export { evaluateAgentPrompt } from './agent-prompt';
+export {
+	evaluateConnections,
+	resolveConnections,
+} from './connections';
+export { evaluateFromAi } from './from-ai';
+export { evaluateTools } from './tools';
+export {
+	evaluateTrigger,
+	type TriggerEvaluationResult,
+} from './trigger';
