+++ packages/@n8n/ai-workflow-builder.ee/src/programmatic/evaluators/tools.ts
@@ -1,0 +1,56 @@
+import type { INodeTypeDescription } from 'n8n-workflow';
+
+import type { SimpleWorkflow } from '@/types';
+
+import type { SingleEvaluatorResult } from '../types';
+import { nodeParametersContainExpression } from '../utils/expressions';
+import { isTool } from '../utils/is-tool';
+import { calcSingleEvaluatorScore } from '../utils/score';
+
+const toolsWithoutParameters = [
+	'@n8n/n8n-nodes-langchain.toolCalculator',
+	'@n8n/n8n-nodes-langchain.toolVectorStore',
+	'@n8n/n8n-nodes-langchain.vectorStoreInMemory',
+	'@n8n/n8n-nodes-langchain.mcpClientTool',
+	'@n8n/n8n-nodes-langchain.toolWikipedia',
+	'@n8n/n8n-nodes-langchain.toolSerpApi',
+];
+
+export function evaluateTools(
+	workflow: SimpleWorkflow,
+	nodeTypes: INodeTypeDescription[],
+): SingleEvaluatorResult {
+	const violations: SingleEvaluatorResult['violations'] = [];
+
+	if (!workflow.nodes || workflow.nodes.length === 0) {
+		return { violations, score: 0 };
+	}
+
+	for (const node of workflow.nodes) {
+		const nodeType = nodeTypes.find((type) => type.name === node.type);
+		if (!nodeType) {
+			continue;
+		}
+
+		if (isTool(nodeType) && !toolsWithoutParameters.includes(node.type)) {
+			if (!node.parameters || Object.keys(node.parameters).length === 0) {
+				violations.push({
+					type: 'major',
+					description: `Tool node "${node.name}" has no parameters set.`,
+					pointsDeducted: 20,
+				});
+				continue;
+			}
+
+			if (!nodeParametersContainExpression(node.parameters)) {
+				violations.push({
+					type: 'major',
+					description: `Tool node "${node.name}" has no expressions in its parameters. This likely means it is not using dynamic input.`,
+					pointsDeducted: 20,
+				});
+			}
+		}
+	}
+
+	return { violations, score: calcSingleEvaluatorScore({ violations }) };
+}
