+++ packages/@n8n/ai-workflow-builder.ee/src/programmatic/utils/is-tool.ts
@@ -1,0 +1,5 @@
+import type { INodeTypeDescription } from 'n8n-workflow';
+
+export function isTool(nodeType: INodeTypeDescription): boolean {
+	return nodeType.codex?.subcategories?.AI?.includes('Tools') ?? false;
+}
