+++ packages/@n8n/ai-workflow-builder.ee/src/utils/test/operations-processor.test.ts
@@ -592,6 +592,7 @@ describe('operations-processor', () => {
 			workflowOperations,
 			messages: [],
 			workflowContext: {},
+			workflowValidation: null,
 			previousSummary: 'EMPTY',
 		});
