+++ packages/@n8n/ai-workflow-builder.ee/src/utils/test/tool-executor.test.ts
@@ -48,6 +48,7 @@ describe('tool-executor', () => {
 			workflowOperations: null,
 			messages,
 			workflowContext: {},
+			workflowValidation: null,
 			previousSummary: 'EMPTY',
 		});
