+++ packages/@n8n/ai-workflow-builder.ee/src/workflow-state.ts
@@ -2,6 +2,8 @@ import type { BaseMessage } from '@langchain/core/messages';
 import { HumanMessage } from '@langchain/core/messages';
 import { Annotation, messagesStateReducer } from '@langchain/langgraph';
 
+import type { ProgrammaticEvaluationResult } from './programmatic/types';
+
 import type { SimpleWorkflow, WorkflowOperation } from './types/workflow';
 import type { ChatPayload } from './workflow-builder-agent';
 
@@ -78,6 +80,10 @@ export const WorkflowState = Annotation.Root({
 	// Latest workflow context
 	workflowContext: Annotation<ChatPayload['workflowContext'] | undefined>({
 		reducer: (x, y) => y ?? x,
+	}),
+	workflowValidation: Annotation<ProgrammaticEvaluationResult | null>({
+		reducer: (x, y) => (y === undefined ? x : y),
+		default: () => null,
 	}),
 
 	// Previous conversation summary (used for compressing long conversations)
